separate BlockCache into the (simple) production use case and the (complex) test use case of mocking a blockchain object
